# store-contract/src/main/resources/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Infinia Store Platform API
  version: 1.0.0
  summary: Catalog, publishing, review and delivery API for the Infinia / FengYu ecosystem.
  description: |
    Versioned REST API of the Infinia Store Platform (design §10).

    * Root path `/api/v1`; breaking changes ship as `/api/v2` — never via headers.
    * JSON is camelCase, timestamps are UTC RFC 3339, identifiers are UUIDv7/ULID.
    * Errors use `application/problem+json` with a stable `code`, `traceId` and
      localizable arguments.
    * Lists use cursor pagination; write endpoints accept `Idempotency-Key`;
      concurrent updates use `ETag/If-Match`.
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.txt
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /listings/{namespace}/{slug}:
    get:
      operationId: getListing
      tags: [Listings]
      summary: Listing detail with visible releases
      security:
        - {}
        - bearerAuth: []
      parameters:
        - { name: namespace, in: path, required: true, schema: { type: string } }
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: channel, in: query, schema: { $ref: '#/components/schemas/Channel' } }
      responses:
        '200':
          description: Listing detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingDetail' }
        '404':
          description: Not found
          content:
            application/problem+json:
              schema: { $ref: '#/components/schemas/Problem' }
  /listings/{namespace}/{slug}/ratings:
    get:
      operationId: listRatings
      tags: [Listings]
      summary: Ratings and average score of a listing (anonymous access allowed)
      security:
        - {}
        - bearerAuth: []
      parameters:
        - { name: namespace, in: path, required: true, schema: { type: string } }
        - { name: slug, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: Ratings page
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RatingsPage' }
    put:
      operationId: upsertRating
      tags: [Moderation]
      summary: Create or update the caller's rating (one per user per listing)
      security:
        - bearerAuth: []
      parameters:
        - { name: namespace, in: path, required: true, schema: { type: string } }
        - { name: slug, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpsertRatingRequest' }
      responses:
        '200':
          description: Stored rating
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Rating' }
...
  /resolutions:
    post:
      operationId: resolve
      tags: [Resolution]
      summary: Resolve versions and dependency closure for a client environment
      security:
        - {}
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ResolveRequest' }
      responses:
        '200':
          description: Resolution plan (resolvable=false when dependencies are missing)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ResolveResponse' }
...
  /releases/{releaseId}/download-ticket:
    post:
      operationId: createDownloadTicket
      tags: [Delivery]
      summary: Mint a short-lived, purpose-limited CDN download ticket
      security:
        - {}
        - bearerAuth: []
      parameters:
        - { name: releaseId, in: path, required: true, schema: { type: string } }
        - { name: artifactId, in: query, schema: { type: string }, description: Select a specific artifact; defaults to the best platform match }
      responses:
        '200':
          description: Ticket
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DownloadTicket' }
        '404': { $ref: '#/components/responses/ProblemNotFound' }
  /releases/{releaseId}/checksums.txt:
    get:
      operationId: releaseChecksums
      tags: [Delivery]
      summary: sha256sum-style checksum manifest of a published release's binary artifacts (design §8.3)
      security:
        - {}
      parameters:
        - { name: releaseId, in: path, required: true, schema: { type: string } }
      responses:
        '200':
          description: 'One `<sha256>  <filename>` line per INSTALLER / PORTABLE / PACKAGE artifact, sorted by filename'
          content:
            text/plain:
              schema: { type: string }
        '404': { $ref: '#/components/responses/ProblemNotFound' }
        '409': { $ref: '#/components/responses/ProblemConflict' }
...
  /publisher/listings:
    get:
      operationId: listPublisherListings
      tags: [Publisher]
      summary: Listings owned by the caller's namespaces
      responses:
        '200':
          description: Listings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CatalogItem' }
    post:
      operationId: createListing
      tags: [Publisher]
      summary: Create a listing in a namespace the caller owns
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateListingRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ListingDetail' }
        '403': { $ref: '#/components/responses/ProblemForbidden' }
        '409': { $ref: '#/components/responses/ProblemConflict' }
...
  /admin/listings:
    get:
      operationId: listAllListings
      tags: [Admin]
      summary: Every listing with curation state — PLATFORM_ADMIN
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All listings regardless of visibility
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AdminListing' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    ProblemForbidden:
      description: Forbidden
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ProblemNotFound:
      description: Not found
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ProblemConflict:
      description: Conflict
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
  schemas:
    ListingType:
      type: string
      enum: [APP, PLUGIN, SKILL, MCP, FLOW]
    Channel:
      type: string
      enum: [stable, beta, alpha, nightly, private]
    Platform:
      type: string
      enum: [windows, macos, linux, universal]
    Arch:
      type: string
      enum: [x64, arm64, universal]
    ArtifactKind:
      type: string
      enum: [PACKAGE, INSTALLER, PORTABLE, CHECKSUMS, SIGNATURE, SBOM]
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code: { type: string, description: Stable machine-readable code }
        detail: { type: string }
        traceId: { type: string }
        parameters:
          type: object
          additionalProperties: true
    CatalogItem:
      type: object
      required: [coordinate, type, namespace, slug, name]
      properties:
        coordinate: { type: string, example: 'infinia://plugin/official/markdown' }
        type: { $ref: '#/components/schemas/ListingType' }
        namespace: { type: string }
        slug: { type: string }
        name: { type: string }
        summary: { type: string }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        iconUrl: { type: string, format: uri }
        latestVersion: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        downloads: { type: integer, format: int64 }
        publisherName: { type: string }
        updatedAt: { type: string, format: date-time }
        featured: { type: boolean, description: Editorial featuring (admin) }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4, description: 'Infinia Level gate: 0 = public; 1..4 restricts view/download to accounts at or above that level' }
...
    Localization:
      type: object
      properties:
        locale: { type: string, example: en }
        name: { type: string }
        summary: { type: string }
        descriptionMarkdown: { type: string }
        changelogMarkdown: { type: string }
    Artifact:
      type: object
      required: [artifactId, kind, platform, arch, filename, size]
      properties:
        artifactId: { type: string }
        kind: { $ref: '#/components/schemas/ArtifactKind' }
        platform: { $ref: '#/components/schemas/Platform' }
        arch: { $ref: '#/components/schemas/Arch' }
        variant: { type: string, description: APP build variant (lite, jre, uos, jar); default for non-APP artifacts }
        filename: { type: string }
        size: { type: integer, format: int64 }
        sha256: { type: string, nullable: true, description: Absent for live upstream delivery }
        keyId: { type: string }
        mimeType: { type: string }
    Dependency:
      type: object
      properties:
        coordinate: { type: string }
        range: { type: string }
        optional: { type: boolean }
    Permission:
      type: object
      properties:
        permissionId: { type: string }
        scope: { type: string, example: network }
        required: { type: boolean }
        reason: { type: string }
    ListingRelease:
      type: object
      required: [releaseId, version, status, channel]
      properties:
        releaseId: { type: string }
        version: { type: string }
        status: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        publishedAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time }
        requiresHost: { type: string }
        license: { type: string }
        sourceUrl: { type: string }
        changelogMarkdown: { type: string }
        rolloutPercent: { type: integer, minimum: 0, maximum: 100 }
        artifacts: { type: array, items: { $ref: '#/components/schemas/Artifact' } }
        dependencies: { type: array, items: { $ref: '#/components/schemas/Dependency' } }
        permissions: { type: array, items: { $ref: '#/components/schemas/Permission' } }
    ListingDetail:
      type: object
      required: [listingId, coordinate, type, namespace, slug]
      properties:
        listingId: { type: string }
        coordinate: { type: string }
        type: { $ref: '#/components/schemas/ListingType' }
        namespace: { type: string }
        slug: { type: string }
        visibility: { type: string, enum: [PUBLIC, UNLISTED, PRIVATE] }
        status: { type: string }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        iconUrl: { type: string, format: uri, nullable: true }
        screenshots: { type: array, items: { type: string, format: uri } }
        defaultChannel: { $ref: '#/components/schemas/Channel' }
        publisherName: { type: string }
        downloads: { type: integer, format: int64 }
        favorites: { type: integer, format: int64 }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4, description: 'Infinia Level gate: 0 = public; 1..4 restricts view/download to accounts at or above that level' }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        localizations: { type: array, items: { $ref: '#/components/schemas/Localization' } }
        releases: { type: array, items: { $ref: '#/components/schemas/ListingRelease' } }
        upstream: { $ref: '#/components/schemas/UpstreamProvenance' }
    UpstreamProvenance:
      type: object
      description: Metadata provenance for a live, non-retained upstream artifact.
      properties:
        sourceName: { type: string }
        externalId: { type: string }
        sourceUrl: { type: string, format: uri }
        sourcePath: { type: string }
        ref: { type: string }
        commitSha: { type: string }
        upstreamVersion: { type: string }
        metadataSha256: { type: string }
        firstSeenAt: { type: string, format: date-time }
        lastSeenAt: { type: string, format: date-time }
        deliveryMode: { type: string, enum: [LIVE_NO_RETENTION] }
    ClientEnvironment:
      type: object
      required: [hostVersion, os, arch]
      properties:
        hostVersion: { type: string }
        os: { $ref: '#/components/schemas/Platform' }
        arch: { $ref: '#/components/schemas/Arch' }
        channel: { $ref: '#/components/schemas/Channel' }
        installed:
          type: array
          items:
            type: object
            properties:
              coordinate: { type: string }
              version: { type: string }
    ResolveRequest:
      type: object
      required: [coordinate, client]
      properties:
        coordinate: { type: string, example: 'infinia://flow/summer/mail-digest' }
        range: { type: string, example: '>=1.0.0 <2.0.0' }
        client: { $ref: '#/components/schemas/ClientEnvironment' }
    ResolutionItem:
      type: object
      properties:
        coordinate: { type: string }
        releaseId: { type: string }
        version: { type: string }
        channel: { $ref: '#/components/schemas/Channel' }
        artifacts: { type: array, items: { $ref: '#/components/schemas/Artifact' } }
        permissions: { type: array, items: { $ref: '#/components/schemas/Permission' } }
        requiresHost: { type: string }
        alreadyInstalled: { type: boolean }
        reason: { type: string }
    MissingDependency:
      type: object
      properties:
        coordinate: { type: string }
        range: { type: string }
        reason: { type: string }
    ResolveResponse:
      type: object
      properties:
        resolvable: { type: boolean }
        rootCoordinate: { type: string }
        plan: { type: array, items: { $ref: '#/components/schemas/ResolutionItem' } }
        missing: { type: array, items: { $ref: '#/components/schemas/MissingDependency' } }
        resolvedAt: { type: string, format: date-time }
    DownloadTicket:
      type: object
      required: [releaseId, artifactId, url]
      properties:
        releaseId: { type: string }
        artifactId: { type: string }
        url: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        sha256: { type: string, nullable: true, description: Absent for live upstream delivery }
        signature: { type: string }
        keyId: { type: string }
        size: { type: integer, format: int64 }
...
    CreateListingRequest:
      type: object
      required: [namespace, slug, type, name]
      properties:
        namespace: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}$' }
        slug: { type: string, pattern: '^[a-z0-9][a-z0-9-]{0,62}$' }
        type: { $ref: '#/components/schemas/ListingType' }
        category: { type: string }
        tags: { type: array, items: { type: string } }
        defaultChannel: { $ref: '#/components/schemas/Channel' }
        name: { type: string, minLength: 1, maxLength: 100 }
        summary: { type: string, maxLength: 500 }
        descriptionMarkdown: { type: string, maxLength: 50000 }
        locale: { type: string, default: en }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4, default: 0, description: 'Infinia Level gate the publisher sets: 0 = public; 1..4 = only signed-in accounts at or above that level may view and download' }
...
    Rating:
      type: object
      properties:
        ratingId: { type: string }
        userId: { type: string }
        stars: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        updatedAt: { type: string, format: date-time }
    RatingSummary:
      type: object
      properties:
        count: { type: integer, format: int64 }
        average: { type: number }
    RatingsPage:
      type: object
      properties:
        summary: { $ref: '#/components/schemas/RatingSummary' }
        ratings: { type: array, items: { $ref: '#/components/schemas/Rating' } }
    UpsertRatingRequest:
      type: object
      required: [stars]
      properties:
        stars: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 2000 }
...
    AdminListing:
      type: object
      properties:
        listingId: { type: string }
        coordinate: { type: string }
        name: { type: string }
        type: { type: string }
        status: { type: string }
        visibility: { type: string, enum: [PUBLIC, UNLISTED, PRIVATE] }
        latestVersion: { type: string, nullable: true }
        featured: { type: boolean }
        minBeeLevel: { type: integer, minimum: 0, maximum: 4 }
        downloads: { type: integer, format: int64 }
...